requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
psycopg>=3.1.8
llama-index>=0.7.0
langchain>=0.0.267
langchain-ollama>=0.0.1
langchain-community>=0.0.1
tavily-python>=0.1.7
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
numpy>=1.24.3
python-dateutil>=2.8.2
torch>=2.0.0
transformers>=4.28.1
difflib36>=0.2.3